refactor(router): making router agnostic about making urls
Remove `makeUrl` method from the router and instead exposed lookup to
find a route by it's name, controller handler or route pattern. This
info can be used with the helpers `makeUrl` method to form proper
URLs. This inturn gives the freedom to create routes with `protocol+domain`
or create signed routes