fix: decodeURI unicode values in the URL, including route params and query string
BREAKING CHANGE: The request.parsedUrl property is no longer of type UrlWithStringQuery, since
url.parse method in Node.js has been deprecated. Instead, we use a custom implementation to
decode the URI and split the query string from it.

Router.match method now accepts an additional 3rd argument. Which is a boolean flag to tell if
it should attempt to decode unicode values in the route params

Closes: #105